Time: 72 ms (40.47%) | Memory: 83.1 MB (47.74%) - LeetSync

+++ 2255-minimum-swaps-to-group-all-1s-together-ii/minimum-swaps-to-group-all-1s-together-ii.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     int minSwaps(vector<int>& nums) {
-        long long k = accumulate(nums.begin(), nums.end(), 0);
-        long long n = nums.size();
-        long long cnt = accumulate(nums.begin(), nums.begin() + k, 0);
-        long long mx = cnt;
+        long k = accumulate(nums.begin(), nums.end(), 0);
+        long n = nums.size();
+        long cnt = accumulate(nums.begin(), nums.begin() + k, 0);
+        long mx = cnt;
         for (int i = k; i < n + k; ++i) {
             cnt += nums[i % n] - nums[(i - k + n) % n];
             mx = max(mx, cnt);
